Update authenticate.handlebars
Changed the welcome setup, added hello to h1 because the header cannot be empty. The only difference is that now in the welcoming animation it will display "welcome, hello, hello!" You can change that if you would like by changing the welcomeanimation.js to have an empty "" to display first. Otherwise, it shouldn't change the front end

Added spaces between required